// src/model/shipping.ts
export type ShippingInputDTO = {
    title: string
    descricao: string
    valor: string
    prazo: string
    destino: string
    origem: string
    pontuacao: string
}

export type ShippingInput = {
    id: string
    title: string
    descricao: string
    valor: string
    prazo: string
    destino: string
    origem: string
    pontuacao: string
}